requirements.txt
flask
textblob
google-generativeai
python-dotenv
openai